refactor: drop InMemoryStateStore, move TaskStateSerde to redis example module
- Delete InMemoryStateStore and InMemoryStateStoreTest from scylla-cdc-lib:
  the null-default path already keeps state in-process via ConcurrentHashMap,
  making a separate InMemoryStateStore class redundant.
- Move TaskStateSerde (and its test) from scylla-cdc-lib to the Redis example
  module (com.scylladb.cdc.lib.state.redis), where it is actually used.
  Make the class and its helpers public so custom CDCStateStore implementations
  can reuse the serialization utilities.
- Update CDCStateStore and CDCConsumer Javadoc to remove references to the
  deleted InMemoryStateStore and TaskStateSerde classes.
