+++ examples/scylla-cdc-state-redis/src/main/java/com/scylladb/cdc/lib/state/redis/RedisStateStore.java
@@ -2,7 +2,6 @@ package com.scylladb.cdc.lib.state.redis;
 
 import com.google.common.base.Preconditions;
 import com.scylladb.cdc.lib.CDCStateStore;
-import com.scylladb.cdc.lib.TaskStateSerde;
 import com.scylladb.cdc.model.GenerationId;
 import com.scylladb.cdc.model.TableName;
 import com.scylladb.cdc.model.TaskId;
