+++ examples/scylla-cdc-state-redis/src/test/java/com/scylladb/cdc/lib/state/redis/TaskStateSerdeTest.java
@@ -1,4 +1,4 @@
-package com.scylladb.cdc.lib;
+package com.scylladb.cdc.lib.state.redis;
 
 import com.scylladb.cdc.model.GenerationId;
 import com.scylladb.cdc.model.StreamId;
